instancetype, preference: Extract apply logic
Again InstancetypeMethods and other public functions with callers
outside of the root instancetype package are left unchanged.

Tests are also left in place for now within instancetype_test.go and
will be broken out at the end of this process to ensure we don't
accidentally change behaviour.

For now the signatures of ApplyToVMI and ApplyToVM differ greatly but
should be aligned in the future.
