+++ pkg/instancetype/apply/iothreadpolicy.go
@@ -1,0 +1,45 @@
+/*
+ * This file is part of the KubeVirt project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * Copyright 2024 Red Hat, Inc.
+ *
+ */
+package apply
+
+import (
+	k8sfield "k8s.io/apimachinery/pkg/util/validation/field"
+
+	virtv1 "kubevirt.io/api/core/v1"
+	v1beta1 "kubevirt.io/api/instancetype/v1beta1"
+)
+
+func applyIOThreadPolicy(
+	field *k8sfield.Path,
+	instancetypeSpec *v1beta1.VirtualMachineInstancetypeSpec,
+	vmiSpec *virtv1.VirtualMachineInstanceSpec,
+) Conflicts {
+	if instancetypeSpec.IOThreadsPolicy == nil {
+		return nil
+	}
+
+	if vmiSpec.Domain.IOThreadsPolicy != nil {
+		return Conflicts{field.Child("domain", "ioThreadsPolicy")}
+	}
+
+	instancetypeIOThreadPolicy := *instancetypeSpec.IOThreadsPolicy
+	vmiSpec.Domain.IOThreadsPolicy = &instancetypeIOThreadPolicy
+
+	return nil
+}
