+++ pkg/instancetype/preference/apply/disk.go
@@ -1,0 +1,69 @@
+//nolint:gocyclo
+/*
+ * This file is part of the KubeVirt project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * Copyright 2024 Red Hat, Inc.
+ *
+ */
+package apply
+
+import (
+	virtv1 "kubevirt.io/api/core/v1"
+	v1beta1 "kubevirt.io/api/instancetype/v1beta1"
+
+	"kubevirt.io/kubevirt/pkg/pointer"
+)
+
+func applyDiskPreferences(preferenceSpec *v1beta1.VirtualMachinePreferenceSpec, vmiSpec *virtv1.VirtualMachineInstanceSpec) {
+	for diskIndex := range vmiSpec.Domain.Devices.Disks {
+		vmiDisk := &vmiSpec.Domain.Devices.Disks[diskIndex]
+		// If we don't have a target device defined default to a DiskTarget so we can apply preferences
+		if vmiDisk.DiskDevice.Disk == nil && vmiDisk.DiskDevice.CDRom == nil && vmiDisk.DiskDevice.LUN == nil {
+			vmiDisk.DiskDevice.Disk = &virtv1.DiskTarget{}
+		}
+
+		if vmiDisk.DiskDevice.Disk != nil {
+			if preferenceSpec.Devices.PreferredDiskBus != "" && vmiDisk.DiskDevice.Disk.Bus == "" {
+				vmiDisk.DiskDevice.Disk.Bus = preferenceSpec.Devices.PreferredDiskBus
+			}
+
+			if preferenceSpec.Devices.PreferredDiskBlockSize != nil && vmiDisk.BlockSize == nil {
+				vmiDisk.BlockSize = preferenceSpec.Devices.PreferredDiskBlockSize.DeepCopy()
+			}
+
+			if preferenceSpec.Devices.PreferredDiskCache != "" && vmiDisk.Cache == "" {
+				vmiDisk.Cache = preferenceSpec.Devices.PreferredDiskCache
+			}
+
+			if preferenceSpec.Devices.PreferredDiskIO != "" && vmiDisk.IO == "" {
+				vmiDisk.IO = preferenceSpec.Devices.PreferredDiskIO
+			}
+
+			if preferenceSpec.Devices.PreferredDiskDedicatedIoThread != nil &&
+				vmiDisk.DedicatedIOThread == nil &&
+				vmiDisk.DiskDevice.Disk.Bus == virtv1.DiskBusVirtio {
+				vmiDisk.DedicatedIOThread = pointer.P(*preferenceSpec.Devices.PreferredDiskDedicatedIoThread)
+			}
+		} else if vmiDisk.DiskDevice.CDRom != nil {
+			if preferenceSpec.Devices.PreferredCdromBus != "" && vmiDisk.DiskDevice.CDRom.Bus == "" {
+				vmiDisk.DiskDevice.CDRom.Bus = preferenceSpec.Devices.PreferredCdromBus
+			}
+		} else if vmiDisk.DiskDevice.LUN != nil {
+			if preferenceSpec.Devices.PreferredLunBus != "" && vmiDisk.DiskDevice.LUN.Bus == "" {
+				vmiDisk.DiskDevice.LUN.Bus = preferenceSpec.Devices.PreferredLunBus
+			}
+		}
+	}
+}
